refactor: Extract statements from parser
